fix code

+++ scripts/after_install/windows.ps1
@@ -19,18 +19,20 @@
 
 if ($env:FEATURE_feature_language_python -eq "1") {
   cd $env:INSTALL_PREFIX/bin
-  $py_dir=(Split-Path -Path  ((get-command python).source))
-  cp *dll ${py_dir}
+    $py_dir=(Split-Path -Path  ((get-command python).source))
+    cp *dll ${py_dir}
   cd $__SRC_DIR
-  if ($env:run_test -eq "1") {
-    if ((Test-Path build -PathType Container)) {
-      rm -r -Force build
-    }
-    Invoke-Expression "$env:CBUILD_PYTHON_EXE -m pytest"
-    if ($env:PACKAGE_VERSION -ne "master") {
-      if ($LastExitCode -ne 0) {
-        exit $LastExitCode
+    if ((Test-Path setup.py -PathType Leaf)) {
+      if ($env:run_test -eq "1") {
+        if ((Test-Path build -PathType Container)) {
+          rm -r -Force build
+        }
+        Invoke-Expression "$env:CBUILD_PYTHON_EXE -m pytest"
+          if ($env:PACKAGE_VERSION -ne "master") {
+            if ($LastExitCode -ne 0) {
+              exit $LastExitCode
+            }
+          }
       }
     }
-  }
 }
